run smoke tests with canary_dist
